backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
redis>=5.0.0
httpx>=0.25.0
PyGithub>=1.59.0
PyJWT>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
conductor-python>=1.1.10